requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
joblib>=1.3
matplotlib>=3.7
openpyxl>=3.1
PyYAML>=6.0
streamlit>=1.32
pytest>=8.0
fastapi>=0.115
uvicorn>=0.30
httpx>=0.27
